backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["abdoafage <abdo.afage2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.2.2"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.1.0"
black = "^23.3.0"
pytest = "^7.4.0"
pytest-django = "^4.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
